SAK-13689
This simply fixes the web.xml so the filter is really activated.

But this coupled with a Kernel change to RemoteHostFilter, will enable the following properties:

webservices.log-allowed=true
webservices.log-denied=true
webservices.allow=localhost,127\.0\.0\.1,192\.168\.[0-9.]+,10\.[0-9.]+,172\.1[6-9]\.[0-9.]+,172\.2[0-9]\.[0-9.]+,172\.3[0-1]\.[0-9.]+
# webservices.allow=:empty:
# webservices.deny=localhost,127\.0\.0\.1,192\.168\.[0-9.]+,10\.[0-9.]+,172\.1[6-9]\.[0-9.]+,172\.2[0-9]\.[0-9.]+,172\.3[0-1]\.

This allows the contents of the web.xml allow filter to be overridden via Sakai properties.

The web.xml is still the default for the filter.



git-svn-id: https://source.sakaiproject.org/svn/webservices/trunk@51343 66ffb92e-73f9-0310-93c1-f5514f145a0a